requirements.txt
Flask==0.11.1
image_slicer==0.1.1
numpy==1.11.1
python_resize_image==1.1.10
scipy==0.18.1
scikit_image==0.12.3
Pillow==3.4.2
